moving existing versioned detectors into subdirectory format
