+++ hack/docs/Adding_Detectors_Internal.md
@@ -48,13 +48,16 @@ If you think that something should be included outside of these guidelines, plea
 
 In some instances, services will update their token format, requiring a new regex to properly detect secrets in addition to supporting the previous token format. Accomodating this can be done without adding a net-new detector. [We provide a `Versioner` interface](https://github.com/trufflesecurity/trufflehog/blob/e18cfd5e0af1469a9f05b8d5732bcc94c39da49c/pkg/detectors/detectors.go#L30) that can be implemented.
 
-1. Create a copy of the package and append `_v2` to the package and file names. Ex: `<packagename>/` -> `<packagename>_v2`, `<packagename>.go` -> `<packagename>_v2.go`
+1. Create two new directories `v1` and `v2`. Move the existing detector and tests into `v1`, and add new files to `v2`.
+Ex: `<packagename>/<old_files>` -> `<packagename>/v1/<old_files>`, `<packagename>/v2/<new_files>`
 
 Note: Be sure to update the tests to reference the new secret values in GSM, or the tests will fail.
 
-2. Implement the `Versioner` interface. [GitHub example implementation.](/pkg/detectors/github_old/github_old.go#L22)
+2. Implement the `Versioner` interface. [GitHub example implementation.](/pkg/detectors/github/v1/github_old.go#L22)
 
-3. Proceed from step 3 of [Creating a new Secret Scanner](#creating-a-new-secret-scanner)
+3. Update the existing detector in DefaultDetectors in `/pkg/engine/defaults.go`
+
+4. Proceed from step 3 of [Creating a new Secret Scanner](#creating-a-new-secret-scanner)
 
 ### Creating a new Secret Scanner
 
@@ -82,7 +85,8 @@ Note: Be sure to update the tests to reference the new secret values in GSM, or 
       4. Found and unverified (indeterminately due to an unexpected API response)
       5. Not found
       6. Any false positive cases that you come across
-   5. Create a merge request for review. CI tests must be passing.
+   5. Add your new detector to DefaultDetectors in `/pkg/engine/defaults.go`
+   6. Create a merge request for review. CI tests must be passing.
 
 ## Addendum
