+++ pkg/detectors/github/v1/github_old_test.go
@@ -1,7 +1,7 @@
 //go:build detectors
 // +build detectors
 
-package maxmindlicense
+package github
 
 import (
 	"context"
@@ -10,23 +10,20 @@ import (
 	"time"
 
 	"github.com/kylelemons/godebug/pretty"
-	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
-
 	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
+	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/pb/detectorspb"
 )
 
-func TestMaxMindLicense_FromChunk(t *testing.T) {
+func TestGitHub_FromChunk(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
 	defer cancel()
-	testSecrets, err := common.GetSecret(ctx, "trufflehog-testing", "detectors3")
+	testSecrets, err := common.GetSecret(ctx, "trufflehog-testing", "detectors2")
 	if err != nil {
 		t.Fatalf("could not get test secrets from GCP: %s", err)
 	}
-	secret := testSecrets.MustGetField("MAXMIND_LICENSE")
-	user := testSecrets.MustGetField("MAXMIND_USER")
-	inactiveSecret := testSecrets.MustGetField("MAXMIND_LICENSE_INACTIVE")
-
+	secret := testSecrets.MustGetField("GITHUB_OLD")
+	secretInactive := testSecrets.MustGetField("GITHUB_OLD_INACTIVE")
 	type args struct {
 		ctx    context.Context
 		data   []byte
@@ -44,14 +41,17 @@ func TestMaxMindLicense_FromChunk(t *testing.T) {
 			s:    Scanner{},
 			args: args{
 				ctx:    context.Background(),
-				data:   []byte(fmt.Sprintf("You can find a geoip secret %s within with maxmind user %s", secret, user)),
+				data:   []byte(fmt.Sprintf("You can find a github secret %s within", secret)),
 				verify: true,
 			},
 			want: []detectors.Result{
 				{
-					DetectorType: detectorspb.DetectorType_MaxMindLicense,
-					Redacted:     "510124",
+					DetectorType: detectorspb.DetectorType_Github,
 					Verified:     true,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/github/",
+						"version":        "1",
+					},
 				},
 			},
 			wantErr: false,
@@ -61,24 +61,38 @@ func TestMaxMindLicense_FromChunk(t *testing.T) {
 			s:    Scanner{},
 			args: args{
 				ctx:    context.Background(),
-				data:   []byte(fmt.Sprintf("You can find a maxmind secret %s within with maxmind user %s", inactiveSecret, user)),
+				data:   []byte(fmt.Sprintf("You can find a github secret %s within", secretInactive)),
 				verify: true,
 			},
 			want: []detectors.Result{
 				{
-					DetectorType: detectorspb.DetectorType_MaxMindLicense,
-					Redacted:     "510124",
+					DetectorType: detectorspb.DetectorType_Github,
 					Verified:     false,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/github/",
+						"version":        "1",
+					},
 				},
 			},
 			wantErr: false,
 		},
 		{
-			name: "not found",
+			name: "not found url",
 			s:    Scanner{},
 			args: args{
 				ctx:    context.Background(),
-				data:   []byte("You cannot find the secret within"),
+				data:   []byte("https://raw.github.com/k/d890e8640f20fba3215ba7be8e0ff145aeb8c17c/include/base64.js"),
+				verify: true,
+			},
+			want:    nil,
+			wantErr: false,
+		},
+		{
+			name: "not found ref",
+			s:    Scanner{},
+			args: args{
+				ctx:    context.Background(),
+				data:   []byte("https://github.com/lz4/lz4 @ dccf8826f1d76efcbdc655e63cc04cdbd1123619"),
 				verify: true,
 			},
 			want:    nil,
@@ -90,17 +104,17 @@ func TestMaxMindLicense_FromChunk(t *testing.T) {
 			s := Scanner{}
 			got, err := s.FromData(tt.args.ctx, tt.args.verify, tt.args.data)
 			if (err != nil) != tt.wantErr {
-				t.Errorf("MaxMindLicense.FromData() error = %v, wantErr %v", err, tt.wantErr)
+				t.Errorf("GitHub.FromData() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
 			for i := range got {
 				if len(got[i].Raw) == 0 {
-					t.Fatalf("no raw secret present: \n %+v", got[i])
+					t.Fatal("no raw secret present")
 				}
 				got[i].Raw = nil
 			}
 			if diff := pretty.Compare(got, tt.want); diff != "" {
-				t.Errorf("MaxMindLicense.FromData() %s diff: (-got +want)\n%s", tt.name, diff)
+				t.Errorf("GitHub.FromData() %s diff: (-got +want)\n%s", tt.name, diff)
 			}
 		})
 	}
