+++ pkg/detectors/jiratoken/v2/jiratoken_v2_test.go
@@ -1,7 +1,7 @@
 //go:build detectors
 // +build detectors
 
-package jiratoken_v2
+package jiratoken
 
 import (
 	"context"
@@ -54,6 +54,10 @@ func TestJiraToken_FromChunk(t *testing.T) {
 				{
 					DetectorType: detectorspb.DetectorType_JiraToken,
 					Verified:     true,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/atlassian/",
+						"version":        "2",
+					},
 				},
 			},
 			wantErr:             false,
@@ -71,6 +75,10 @@ func TestJiraToken_FromChunk(t *testing.T) {
 				{
 					DetectorType: detectorspb.DetectorType_JiraToken,
 					Verified:     false,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/atlassian/",
+						"version":        "2",
+					},
 				},
 			},
 			wantErr:             false,
@@ -100,6 +108,10 @@ func TestJiraToken_FromChunk(t *testing.T) {
 				{
 					DetectorType: detectorspb.DetectorType_JiraToken,
 					Verified:     false,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/atlassian/",
+						"version":        "2",
+					},
 				},
 			},
 			wantErr:             false,
@@ -117,6 +129,10 @@ func TestJiraToken_FromChunk(t *testing.T) {
 				{
 					DetectorType: detectorspb.DetectorType_JiraToken,
 					Verified:     false,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/atlassian/",
+						"version":        "2",
+					},
 				},
 			},
 			wantErr:             false,
