+++ pkg/detectors/maxmindlicense/v1/maxmindlicense_test.go
@@ -1,7 +1,7 @@
 //go:build detectors
 // +build detectors
 
-package github_old
+package maxmindlicense
 
 import (
 	"context"
@@ -10,20 +10,23 @@ import (
 	"time"
 
 	"github.com/kylelemons/godebug/pretty"
-	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
+
+	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/pb/detectorspb"
 )
 
-func TestGitHub_FromChunk(t *testing.T) {
+func TestMaxMindLicense_FromChunk(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
 	defer cancel()
-	testSecrets, err := common.GetSecret(ctx, "trufflehog-testing", "detectors2")
+	testSecrets, err := common.GetSecret(ctx, "trufflehog-testing", "detectors3")
 	if err != nil {
 		t.Fatalf("could not get test secrets from GCP: %s", err)
 	}
-	secret := testSecrets.MustGetField("GITHUB_OLD")
-	secretInactive := testSecrets.MustGetField("GITHUB_OLD_INACTIVE")
+	secret := testSecrets.MustGetField("MAXMIND_LICENSE")
+	user := testSecrets.MustGetField("MAXMIND_USER")
+	inactiveSecret := testSecrets.MustGetField("MAXMIND_LICENSE_INACTIVE")
+
 	type args struct {
 		ctx    context.Context
 		data   []byte
@@ -41,13 +44,18 @@ func TestGitHub_FromChunk(t *testing.T) {
 			s:    Scanner{},
 			args: args{
 				ctx:    context.Background(),
-				data:   []byte(fmt.Sprintf("You can find a github secret %s within", secret)),
+				data:   []byte(fmt.Sprintf("You can find a geoip secret %s within with maxmind user %s", secret, user)),
 				verify: true,
 			},
 			want: []detectors.Result{
 				{
-					DetectorType: detectorspb.DetectorType_Github,
+					DetectorType: detectorspb.DetectorType_MaxMindLicense,
+					Redacted:     "510124",
 					Verified:     true,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/maxmind/",
+						"version":        "1",
+					},
 				},
 			},
 			wantErr: false,
@@ -57,34 +65,28 @@ func TestGitHub_FromChunk(t *testing.T) {
 			s:    Scanner{},
 			args: args{
 				ctx:    context.Background(),
-				data:   []byte(fmt.Sprintf("You can find a github secret %s within", secretInactive)),
+				data:   []byte(fmt.Sprintf("You can find a maxmind secret %s within with maxmind user %s", inactiveSecret, user)),
 				verify: true,
 			},
 			want: []detectors.Result{
 				{
-					DetectorType: detectorspb.DetectorType_Github,
+					DetectorType: detectorspb.DetectorType_MaxMindLicense,
+					Redacted:     "510124",
 					Verified:     false,
+					ExtraData: map[string]string{
+						"rotation_guide": "https://howtorotate.com/docs/tutorials/maxmind/",
+						"version":        "1",
+					},
 				},
 			},
 			wantErr: false,
 		},
 		{
-			name: "not found url",
+			name: "not found",
 			s:    Scanner{},
 			args: args{
 				ctx:    context.Background(),
-				data:   []byte("https://raw.github.com/k/d890e8640f20fba3215ba7be8e0ff145aeb8c17c/include/base64.js"),
-				verify: true,
-			},
-			want:    nil,
-			wantErr: false,
-		},
-		{
-			name: "not found ref",
-			s:    Scanner{},
-			args: args{
-				ctx:    context.Background(),
-				data:   []byte("https://github.com/lz4/lz4 @ dccf8826f1d76efcbdc655e63cc04cdbd1123619"),
+				data:   []byte("You cannot find the secret within"),
 				verify: true,
 			},
 			want:    nil,
@@ -96,17 +98,17 @@ func TestGitHub_FromChunk(t *testing.T) {
 			s := Scanner{}
 			got, err := s.FromData(tt.args.ctx, tt.args.verify, tt.args.data)
 			if (err != nil) != tt.wantErr {
-				t.Errorf("GitHub.FromData() error = %v, wantErr %v", err, tt.wantErr)
+				t.Errorf("MaxMindLicense.FromData() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
 			for i := range got {
 				if len(got[i].Raw) == 0 {
-					t.Fatal("no raw secret present")
+					t.Fatalf("no raw secret present: \n %+v", got[i])
 				}
 				got[i].Raw = nil
 			}
 			if diff := pretty.Compare(got, tt.want); diff != "" {
-				t.Errorf("GitHub.FromData() %s diff: (-got +want)\n%s", tt.name, diff)
+				t.Errorf("MaxMindLicense.FromData() %s diff: (-got +want)\n%s", tt.name, diff)
 			}
 		})
 	}
